Issue #1993 New returned mail case note script
This script will case note the addr the returned mail came from, the forwarding addr if there is one, the county the new addr is in, if the returned mail was resent, if additional forms were sent, if the case is known on MNsure, the MNsure number if on MNsure, and if the addr is correct on MNsure. It will also set an auto TIKL if the 2919A was sent but will pause to allow the worker to review the case note prior to doing the auto TIKL. Any "other notes/actions taken" will be case noted as well.If the addr is incorrect on MNsure the worker will get a message box pop-up to update if possible or to forward the info to the appropriate area to be updated.